Change: fix google chrome add repo and add fancy_echo

+++ installer/ubuntu.sh
@@ -5,9 +5,15 @@
 # works (https://github.com/thoughtbot/laptop)
 
 # how to run
-# curl --remote-name [change to raw file url] && sudo chmod +x ubuntu.sh && ./ubuntu.sh 2>&1 | tee ~/laptop.log
+# curl --remote-name https://raw.githubusercontent.com/andymarthin/dotfiles/master/installer/ubuntu.sh && sudo chmod +x ubuntu.sh && ./ubuntu.sh 2>&1 | tee ~/laptop.log
 
-sudo apt update && sudo apt upgrade -y
+fancy_echo() {
+  local fmt="$1"; shift
+
+  # shellcheck disable=SC2059
+  printf "\\n$fmt\\n" "$@"
+}
+
 append_to_zshrc() {
   if [ -w "$HOME/.zshrc.local" ]; then
     zshrc="$HOME/.zshrc.local"
@@ -25,36 +31,36 @@ set -e
 
 append_to_zshrc 'export PATH="$HOME/.bin:$PATH"'
 
-# add all ppa's first
-#sudo add-apt-repository -y "deb https://cli-assets.heroku.com/branches/stable/apt ./"
-# sudo add-apt-repository -y ppa:martin-frost/thoughtbot-rcm
-# curl -sL https://deb.nodesource.com/setup_11.x | sudo -E bash -
-# sudo apt-get install -y nodejs
+sudo apt update && sudo apt upgrade -y
 
 # basics
+fancy_echo "Installing libraries for common gem dependencies ..."
 sudo apt-get install -y libssl-dev libreadline-dev zlib1g-dev autoconf bison build-essential libyaml-dev libreadline-dev libncurses5-dev libffi-dev libgdbm-dev snapd ctags git tmux vim zsh wget
-chsh -s "$(which zsh)"
+sudo chsh -s "$(which zsh)"
+
+fancy_echo "Installing oh my zsh ..."
 # oh my zsh
 if [ ! -e "$HOME/.oh-my-zsh" ]; then
   sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
 fi
+
+fancy_echo "Setup git config global user name ..."
 if [ -z "$(git config --global user.name)" ]; then
   echo Your Name?
   read name
   git config --global user.name $name
 fi
 
+fancy_echo "Setup git config globa user email ..."
 if [ -z "$(git config --global user.email)"  ]; then
   echo Your Email?
   read email
   git config --global user.email $email
 fi
 
-
-# heroku
-#sudo apt-get install -y heroku
+#Heroku
+fancy_echo "Installng heroku"
 sudo snap install --classic heroku
-
 
 # image manip
 sudo apt-get install -y imagemagick libmagickwand-dev --fix-missing
@@ -62,28 +68,35 @@ sudo apt-get install -y imagemagick libmagickwand-dev --fix-missing
 # capybara-webkit dependencies
 sudo apt-get install -y qt5-default libqt5webkit5-dev gstreamer1.0-plugins-base gstreamer1.0-tools gstreamer1.0-x
 
-
-# sudo apt-get install -y nodejs
-sudo snap install node --channel=10/stable --classic
-
+# install nodejs
+fancy_echo "Installing nodejs ..."
+sudo snap install node --channel=12/stable --classic
 
 # databases
+fancy_echo "Installing Redis ..."
 sudo apt-get install -y redis-server
+fancy_echo "Installing Postgresql ..."
 sudo apt-get install -y postgresql postgresql-contrib libpq-dev
 
 # visual code
+fancy_echo "Installing Vscode ..."
 sudo snap install code --classic
 
 # POSTMAN
+fancy_echo "Installing Postman ..."
 sudo snap install postman
 
 # INSTALL GOOGLE CHROME
-wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
-sudo sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list'
-sudo apt update
-sudo apt install -y google-chrome-stable
+fancy_echo "Installing Google Chrome ..."
+if [ ! -e "/etc/apt/sources.list.d/google-chrome.list" ]; then
+  wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
+  sudo sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list'
+  sudo apt update
+fi
+[ ! -z "$(which google-chrome)" ] && sudo apt install -y google-chrome-stable
 
-# language related
+# Install Rbenv
+fancy_echo "Installing Rbenv ..."
 [ ! -e "$HOME/.rbenv" ] && git clone https://github.com/rbenv/rbenv.git ~/.rbenv
 [ ! -e "$HOME/.rbenv/plugins/ruby-build" ] && git clone https://github.com/rbenv/ruby-build.git ~/.rbenv/plugins/ruby-build
 if [ -z "$(which rbenv)" ]; then
@@ -96,7 +109,6 @@ fi
 append_to_zshrc 'export PATH="$HOME/.rbenv/bin:$PATH"'
 append_to_zshrc 'eval "$(rbenv init - --no-rehash)"'
 
-
 find_latest_ruby() {
   rbenv install -l | grep -v - | tail -1 | sed -e 's/^ *//'
 }
@@ -106,6 +118,7 @@ eval "$(rbenv init -)"
 rbenv install -s "$ruby_version"
 
 if ! rbenv versions | grep -Fq "$ruby_version"; then
+  fancy_echo "Instaling latesest Ruby ..."
   rbenv install -s "$ruby_version"
 fi
 
@@ -115,3 +128,4 @@ rbenv shell "$ruby_version"
 gem update --system
 gem install bundler
 gem install parity
+gem install rails
